+++ Frontend/Models/Benefits/OtherFactorsItemViewModel.cs
@@ -1,0 +1,11 @@
+﻿using Data.Models.Projects;
+
+namespace Frontend.Models.Benefits
+{
+    public class OtherFactorsItemViewModel
+    {
+        public TransferBenefits.OtherFactor OtherFactor { get; set; }
+        public bool Checked { get; set; }
+        public string Description { get; set; }
+    }
+}
